refactor: Simplify NFT gating and enhance contract structure
- Updated import statements across multiple contracts for consistency and clarity.
- Refactored access control logic in `CookieJar` and `CookieJarFactory` to streamline NFT gating, transitioning to a unified `NFTRequirement` structure.
- Removed obsolete libraries and functions related to NFT validation and access initialization, reducing complexity.
- Enhanced the `AccessConfig` struct to support simplified access management.
- Updated tests to reflect changes in access control and ensure comprehensive coverage of new functionalities.

These changes improve the maintainability of the codebase and enhance the overall functionality of the Cookie Jar protocol.

+++ contracts/script/Deploy.s.sol
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.24;
 
-import "forge-std/Script.sol";
-import "../src/CookieJarFactory.sol";
+import {Script, stdJson, console} from "forge-std/Script.sol";
+import {CookieJarFactory} from "../src/CookieJarFactory.sol";
 
 /// @title Universal Deploy Script with Auto-Config Updates
 /// @notice Deploys contracts and automatically updates client configuration
@@ -33,10 +33,10 @@ contract Deploy is Script {
         
         CookieJarFactory factory = new CookieJarFactory(
             feeCollector,
-            owner, 
+            owner,
             feePercentage,
-            minETHDeposit,
-            minERC20Deposit
+            uint128(minETHDeposit),
+            uint128(minERC20Deposit)
         );
         
         vm.stopBroadcast();
